разбивает программу на модули
