+++ js/util.js
@@ -1,0 +1,28 @@
+function getRandomInt(min, max) {
+  if (min <= 0 || max <= 0 || max <= min) {
+    return NaN;
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min; //Максимум и минимум включаются
+}
+
+function getRandomPositiveInteger (a, b) {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+const checkLengthString = (string, maxLength) => string.length <= maxLength;
+
+checkLengthString();
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {getRandomInt};
+export {getRandomPositiveInteger};
+export {getRandomArrayElement};
